+++ hive_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Adds `IsolatedHive.initFlutter`
+
 ## 2.2.0
 
 - Fixes deprecation of `Color.value`